move model builder to automl base model
