+++ pyzoo/zoo/automl/model/__init__.py
@@ -13,4 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from .model_builder import *
