+++ pyzoo/zoo/automl/model/base_pytorch_model.py
@@ -18,8 +18,7 @@ from torch.utils.data import TensorDataset, DataLoader
 
 import types
 
-from zoo.automl.model import ModelBuilder
-from zoo.automl.model.abstract import BaseModel
+from zoo.automl.model.abstract import BaseModel, ModelBuilder
 from zoo.automl.common.util import *
 from zoo.automl.common.metrics import Evaluator
 import pandas as pd
