+++ pyzoo/zoo/chronos/model/time_sequence.py
@@ -15,8 +15,7 @@
 #
 from copy import deepcopy
 
-from zoo.automl.model import ModelBuilder
-from zoo.automl.model.abstract import BaseModel
+from zoo.automl.model.abstract import BaseModel, ModelBuilder
 from zoo.chronos.model.VanillaLSTM import VanillaLSTM
 from zoo.chronos.model.Seq2Seq import LSTMSeq2Seq
 from zoo.chronos.model.MTNet_keras import MTNetKeras
