+++ pyzoo/zoo/orca/automl/xgboost/XGBoost.py
@@ -19,8 +19,7 @@ from xgboost.sklearn import XGBRegressor
 
 from xgboost.sklearn import XGBClassifier
 from zoo.automl.common.metrics import Evaluator
-from zoo.automl.model import ModelBuilder
-from zoo.automl.model.abstract import BaseModel
+from zoo.automl.model.abstract import BaseModel, ModelBuilder
 import logging
 
 logger = logging.getLogger(__name__)
